+++ library/src/main/java/io/ona/kujaku/utils/IOUtil.java
@@ -1,0 +1,28 @@
+package io.ona.kujaku.utils;
+
+import java.io.BufferedInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+
+/**
+ * @author Vincent Karuri
+ */
+public class IOUtil {
+
+    public  static final String TAG = IOUtil.class.getName();
+
+    public static String readInputStreamAsString(InputStream in)
+            throws IOException {
+
+        BufferedInputStream bis = new BufferedInputStream(in);
+        ByteArrayOutputStream buf = new ByteArrayOutputStream();
+        int result = bis.read();
+        while(result != -1) {
+            byte b = (byte)result;
+            buf.write(b);
+            result = bis.read();
+        }
+        return buf.toString();
+    }
+}
