+++ utils/src/main/java/io/ona/kujaku/utils/Constants.java
@@ -33,5 +33,4 @@ public class Constants {
 
     public static final String INSERT_OR_REPLACE = "INSERT OR REPLACE INTO %s VALUES ";
     public static final String DATABASE_NAME = "kujaku.db";
-
 }
